# docs/swagger.yaml
basePath: /
definitions:
...
  admin.SaveAdminVideoRequest:
    properties:
      ad_template_id:
        type: string
      description:
        type: string
      duration:
        type: integer
      format:
        type: string
      size:
        type: integer
      status:
        type: string
      title:
        type: string
      url:
        type: string
      user_id:
        type: string
    required:
    - size
    - title
    - url
    - user_id
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for Stream application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stream API
  version: "1.0"
paths:
...
  /admin/videos:
    get:
      description: Get paginated list of all videos across users (admin only)
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Search by title
        in: query
        name: search
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List All Videos
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a model video record for a user (admin only)
      parameters:
      - description: Video payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.SaveAdminVideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Video
      tags:
      - admin
  /admin/videos/{id}:
    delete:
      description: Delete any video by ID (admin only)
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Video (Admin)
      tags:
      - admin
    get:
      description: Get video detail by ID (admin only)
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Video Detail
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update video metadata and status (admin only)
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      - description: Video payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.SaveAdminVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update Video
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"